Release 1.0.0: remove ontology import

+++ tests/compiler_tests/duckdb/20_list_comprehension.sql
@@ -12,7 +12,7 @@ drop type if exists logicarecord481217614 cascade; create type logicarecord48121
 drop type if exists logicarecord383307722 cascade; create type logicarecord383307722 as struct(a timestamp);
 
 -- Logica type: logicarecord884343024
-drop type if exists logicarecord884343024 cascade; create type logicarecord884343024 as struct(arg double, value double);
+drop type if exists logicarecord884343024 cascade; create type logicarecord884343024 as struct(arg numeric, value numeric);
 
 -- Logica type: logicarecord519939597
 drop type if exists logicarecord519939597 cascade; create type logicarecord519939597 as struct(args text[], predicate text);
